Analytics: Move cart functions out of the library
