+++ client/lib/analytics/cart.js
@@ -8,7 +8,9 @@ import { differenceWith, get, isEqual, each, omit } from 'lodash';
  * Internal dependencies
  */
 import analytics from 'lib/analytics';
+import { costToUSD } from 'lib/analytics/utils';
 import { getAllCartItems } from 'lib/cart-values/cart-items';
+import { recordAddToCart as recordAddToCartTracking, recordOrder } from 'lib/analytics/ad-tracking';
 
 export function recordEvents( previousCart, nextCart ) {
 	const previousItems = getAllCartItems( previousCart );
@@ -40,4 +42,33 @@ export function recordUnrecognizedPaymentMethod( action ) {
 	};
 
 	analytics.tracks.recordEvent( 'calypso_cart_unrecognized_payment_method', eventArgs );
+}
+
+export function recordAddToCart( { cartItem } ) {
+	// TODO: move Tracks event here?
+	// Google Analytics
+	const usdValue = costToUSD( cartItem.cost, cartItem.currency );
+	analytics.ga.recordEvent(
+		'Checkout',
+		'calypso_cart_product_add',
+		'',
+		usdValue ? usdValue : undefined
+	);
+	// Marketing
+	recordAddToCartTracking( cartItem );
+}
+
+export function recordPurchase( { cart, orderId } ) {
+	if ( cart.total_cost >= 0.01 ) {
+		// Google Analytics
+		const usdValue = costToUSD( cart.total_cost, cart.currency );
+		analytics.ga.recordEvent(
+			'Purchase',
+			'calypso_checkout_payment_success',
+			'',
+			usdValue ? usdValue : undefined
+		);
+		// Marketing
+		recordOrder( cart, orderId );
+	}
 }
